+++ ProtocolProcessing/simulation.c
@@ -2,6 +2,7 @@
 #define WIN32_LEAN_AND_MEAN
 #include <windows.h>
 
+#include <malloc.h>
 #include <memory.h>
 
 #include "simulation.h"
@@ -15,54 +16,17 @@
 #endif
 
 
-
-
-typedef Int (*socket_read_func)(Pointer,Int8*,Int);
-typedef Int (*socket_write_func)(Pointer,Int8*,Int);
-
-typedef Int (*socket_connect_func)(Pointer,network_addr);
-
 typedef struct{
-	
-	simulation_socket_type m_type;
 
-	Pointer m_in;
-	Pointer m_out;
+	network_area m_area;
+	network_node* m_gateway;
+	Pointer m_node_list;
+} SIMULATION_NETWORK;
 
-	socket_read_func m_read;
-	socket_write_func m_write;
+static Pointer m_network_list = 0;
 
-	socket_connect_func m_connect;
+#define PSIMULATION_NETWORK(network) ((SIMULATION_NETWORK*)network)
 
-} SOCKET;
-
-typedef struct{
-	SOCKET m_socket;
-
-	Int m_port;
-	Int m_listen;
-
-} BOUND_SOCKET;
-
-static network_node *parent_node = 0;
-
-#define PSOCKET(socket) ((SOCKET*)socket)
-
-Int simulation_write_raw_socket(Pointer socket, Int8* data, Int size){
-	
-	UInt32 bytesWritten = 0;
-	WriteFile(PSOCKET(socket)->m_in,data,size,&bytesWritten,0);
-	
-	return bytesWritten;
-}
-
-Int simulation_read_raw_socket(Pointer socket,Int8* data , Int size){
-
-	UInt32 bytesRead = 0;
-	ReadFile(PSOCKET(socket)->m_out,data,size,&bytesRead,0);
-
-	return bytesRead;
-}
 // TODO: Add BGP support.
 network_node* simulation_find_peer(network_node* node , network_addr addr){
 	
@@ -82,45 +46,24 @@ network_node* simulation_find_peer(network_node* node , network_addr addr){
 	return 0;
 }
 
-Int simulation_connect_raw_socket(Pointer socket, network_addr addr){
-	
-	Pointer peer = simulation_find_peer(parent_node,addr);
 
-	if(!peer){
+Pointer simulation_create_network(network_area area){
 
+	SIMULATION_NETWORK* simulation_network = malloc(sizeof(SIMULATION_NETWORK));
+
+	simulation_network->m_area = area;
+	simulation_network->m_gateway = network_create_node();
+	memcpy(simulation_network->m_gateway->address,area.gateway,4);
+
+	simulation_network->m_node_list = list_create();
+
+	//list_add_item(m_network_list,simulation_network);
+	return simulation_network;
+}
+
+
+void simulation_start(void){
+	if(!m_network_list){
+		m_network_list = list_create();
 	}
-
-	return 0;
-}
-
-Int simulation_read_socket(Pointer socket,Int8* data , Int size){
-	
-	return PSOCKET(socket)->m_read(socket,data,size);
-}
-
-Int simulation_write_socket(Pointer socket,Int8* data , Int size){
-	
-	return PSOCKET(socket)->m_write(socket,data,size);
-}
-
-
-Pointer simulation_socket(simulation_socket_type type){
-	
-	SOCKET* socket = malloc(sizeof(SOCKET));
-	memset(socket,0,sizeof(SOCKET));
-	
-	CreatePipe(&socket->m_out,&socket->m_in,0,0);
-
-	socket->m_read = simulation_read_raw_socket;
-	socket->m_write = simulation_write_raw_socket;
-
-	socket->m_connect = simulation_connect_raw_socket;
-
-	socket->m_type = type;
-
-	return socket;
-}
-
-Int simulation_connect(Pointer socket, network_addr addr){
-	PSOCKET(socket)->m_connect(socket,addr);
 }
